refactor(presets): maintain preset codes inside `src`
