+++ presets/index.d.ts
@@ -1,17 +1,3 @@
-import type { TsJestPresets } from '../dist'
+import allPresets from '../dist/presets/all-presets'
 
-declare const _default: {
-    defaults: TsJestPresets;
-    defaultsLegacy: TsJestPresets;
-    defaultsESM: TsJestPresets;
-    defaultsESMLegacy: TsJestPresets;
-    jsWithTs: TsJestPresets;
-    jsWithTsLegacy: TsJestPresets;
-    jsWithTsESM: TsJestPresets;
-    jsWithTsESMLegacy: TsJestPresets;
-    jsWithBabel: TsJestPresets;
-    jsWithBabelLegacy: TsJestPresets;
-    jsWithBabelESM: TsJestPresets;
-    jsWithBabelESMLegacy: TsJestPresets;
-};
-export = _default;
+export = allPresets
